chore: follow authors patterns of hooks usage
