+++ src/renderer/src/App.tsx
@@ -1,5 +1,4 @@
-import { useEffect, useState, useRef, KeyboardEvent } from 'react'
-import { AlertTriangle } from 'lucide-react'
+import { useState, useRef, KeyboardEvent } from 'react'
 
 import { Command, CommandInput, CommandList } from '@renderer/elements/Command'
 import { CommandEmpty, CommandShortcut } from '@renderer/elements/Command'
@@ -8,10 +7,9 @@ import { Commands } from '@renderer/components/Commands'
 import { CommandPage } from '@renderer/components/CommandPage'
 import { CommandApplications } from '@renderer/components/CommandApplications'
 import { CommandShortcuts } from '@renderer/components/CommandShortcuts'
-import { useScrollToTop } from '@renderer/hooks'
+import { usePluginToasts, useScrollToTop } from '@renderer/hooks'
 import { Settings } from '@renderer/components/Settings'
-import { winElectron } from '@renderer/lib/utils'
-import { Toaster, toast as sonnerToast } from 'sonner'
+import { Toaster } from 'sonner'
 
 const App = () => {
   const [selectedCommand, setSelectedCommand] = useState<CommandT | null>(null)
@@ -19,38 +17,9 @@ const App = () => {
   const commandListRef = useRef<HTMLDivElement | null>(null)
   const [currentBangName, setCurrentBangName] = useState<string | null>(null)
 
-  useEffect(() => {
-    if (!winElectron?.onPluginToast) return
-
-    const unsubscribe = winElectron.onPluginToast(({ type, title, description, duration }) => {
-      const toastOptions = { description, duration }
-
-      switch (type) {
-        case 'success':
-          sonnerToast.success(title, toastOptions)
-          break
-        case 'warning':
-          sonnerToast.warning(title, toastOptions)
-          break
-        case 'error':
-          sonnerToast.error(title, {
-            ...toastOptions,
-            icon: <AlertTriangle  />,
-            className: '!gap-4',
-          })
-          break
-        default:
-          sonnerToast(title, toastOptions)
-          break
-      }
-    })
-
-    return () => {
-      if (typeof unsubscribe === 'function') unsubscribe()
-    }
-  }, [])
-
+  usePluginToasts()
   useScrollToTop(commandListRef, [commandSearch])
+
   const handleInputKeyDown = (e: KeyboardEvent<HTMLInputElement>) => {
     if (e.key === 'Escape') {
       e.preventDefault()
@@ -68,7 +37,8 @@ const App = () => {
 
   return (
     <div className="bg-black h-full">
-      <Toaster position="bottom-right" theme="dark" richColors  />
+      <Toaster position="bottom-right" theme="dark" richColors />
+
       {!selectedCommand && (
         <Command filter={commandFilter} loop>
           <div className="flex items-center gap-2 px-3 border-b border-zinc-800">
